+++ src/promptflow/promptflow/__init__.py
@@ -13,7 +13,6 @@ from promptflow._core.tracer import trace
 from promptflow._sdk._load_functions import load_flow, load_run
 
 from ._sdk._pf_client import PFClient
-from ._trace._start_trace import start_trace
 from ._version import VERSION
 
 # backward compatibility
@@ -21,4 +20,4 @@ log_flow_metric = log_metric
 
 __version__ = VERSION
 
-__all__ = ["PFClient", "load_flow", "load_run", "log_metric", "start_trace", "ToolProvider", "tool", "trace"]
+__all__ = ["PFClient", "load_flow", "load_run", "log_metric", "ToolProvider", "tool", "trace"]
