+++ src/promptflow/promptflow/executor/_line_execution_process_pool.py
@@ -667,7 +667,7 @@ def _process_wrapper(
     OperationContext.get_instance().update(operation_contexts_dict)  # Update the operation context for the new process.
 
     # set up OpenTelemetry exporter in process who executes the line
-    from promptflow._trace._start_trace import setup_exporter_from_environ
+    from promptflow.tracing._start_trace import setup_exporter_from_environ
 
     setup_exporter_from_environ()
