+++ src/promptflow/promptflow/tracing/__init__.py
@@ -3,3 +3,7 @@
 # ---------------------------------------------------------
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+
+from ._start_trace import start_trace
+
+__all__ = ["start_trace"]
